remove eduid-common dependency
